refactor out chrome mock
